+++ tools/cli/vz/cmd/logout/logout.go
@@ -4,15 +4,11 @@
 package logout
 
 import (
-	"errors"
 	"fmt"
 	"github.com/spf13/cobra"
+	"github.com/verrazzano/verrazzano/tools/cli/vz/pkg/helpers"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	"k8s.io/client-go/tools/clientcmd"
-	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
-	"k8s.io/client-go/util/homedir"
-	"os"
-	"path/filepath"
 	"strings"
 )
 
@@ -48,20 +44,36 @@ func NewCmdLogout(streams genericclioptions.IOStreams) *cobra.Command {
 }
 
 func logout() error {
-	kubeConfigLoc,err := getKubeconfigLocation()
+
+	// Obtain the default kubeconfig's location
+	kubeConfigLoc,err := helpers.GetKubeconfigLocation()
 	if err!=nil {
 		return err
 	}
+
+	// Load the default kubeconfig's configuration into clientcmdapi object
 	mykubeConfig, _ := clientcmd.LoadFromFile(kubeConfigLoc)
+	if err!=nil {
+		fmt.Println("Unable to load kubeconfig, check permissions")
+		return err
+	}
 
-	delete(mykubeConfig.Clusters, "verrazzano")
-	delete(mykubeConfig.AuthInfos,"verrazzano")
-	delete(mykubeConfig.Contexts,mykubeConfig.CurrentContext)
-	mykubeConfig.CurrentContext = strings.Split(mykubeConfig.CurrentContext,"@")[1]
+	// Remove the cluster with nickname verrazzano
+	helpers.RemoveCluster(mykubeConfig, "verrazzano")
 
-	err = WriteKubeConfigToDisk( kubeConfigLoc,
-		mykubeConfig,
-	)
+	// Remove the user with nickname verrazzano
+	helpers.RemoveUser(mykubeConfig,"verrazzano")
+
+	// Remove the currentcontext
+	helpers.RemoveContext(mykubeConfig, mykubeConfig.CurrentContext)
+
+	// Set currentcluster to the cluster before the user logged in
+	helpers.SetCurrentContext(mykubeConfig, strings.Split(mykubeConfig.CurrentContext,"@")[1])
+
+	// Write kubeconfig to file
+	err = clientcmd.WriteToFile(*mykubeConfig,
+								kubeConfigLoc,
+								)
 	if err!=nil {
 		fmt.Println("Unable to write the new kubconfig to disk")
 		return err
@@ -70,30 +82,4 @@ func logout() error {
 	return nil
 }
 
-// Write the kubeconfig object to a file in yaml format
-func WriteKubeConfigToDisk(filename string, kubeconfig *clientcmdapi.Config) error {
-	err := clientcmd.WriteToFile(*kubeconfig, filename)
-	if err != nil {
-		return err
-	}
-	return nil
-}
 
-// Helper function to obtain the default kubeconfig location
-func getKubeconfigLocation() (string,error) {
-
-	var kubeconfig string
-	kubeconfigEnvVar := os.Getenv("KUBECONFIG")
-
-	if len(kubeconfigEnvVar) > 0 {
-		// Find using environment variables
-		kubeconfig = kubeconfigEnvVar
-	} else if home := homedir.HomeDir(); home != "" {
-		// Find in the ~/.kube/ directory
-		kubeconfig = filepath.Join(home, ".kube", "config")
-	} else {
-		// give up
-		return "", errors.New("Could not find kube config")
-	}
-	return kubeconfig,nil
-}
