+++ tools/cli/vz/pkg/helpers/browser.go
@@ -1,0 +1,23 @@
+package helpers
+
+import (
+	"fmt"
+	"os/exec"
+	"runtime"
+)
+
+// Non-blocking browser opener function
+func OpenUrlInBrowser(url string) error{
+	var err error
+	switch runtime.GOOS {
+	case "linux":
+		err = exec.Command("xdg-open", url).Start()
+	case "windows":
+		err = exec.Command("rundll32", "url.dll,FileProtocolHandler", url).Start()
+	case "darwin":
+		err = exec.Command("open", url).Start()
+	default:
+		err = fmt.Errorf("unsupported platform")
+	}
+	return err
+}
